Generators for condition blocks
